fix: after >1 guesses resizing the window would play audio
This issue was caused by two factors:
1. A misunderstanding about continue in a do-while
2. Not initializing guess to a known value resulting in UB

I wrongly thought continue in a do-while went to the top of the loop,
however it goes to the condition at the bottom. This meant the loop
condition was being tested when I didn't think it was.

Not initializing a variable can lead to UB, but because I thought
continue was going to the top of the loop I didn't believe the value of
guess was being used and therefore didn't think UB would occur. In my
mind, by the time the value of guess was used it would be initialized by
the sanitization method under the switch.