feat: restructure the morse_table to be a more opaque type
This wasn't strictly necessary and was done to explore opaque types and
function pointers. However, it did also seem nice to be able to restrict
access to the table to specific functions and prevent the handful of
places that use the table from getting too creative. OOP habits die hard.